Fix navbar dropdown menu for active item
Due to the upstream changes in #10 the custom Trustlines colorscheme
does not work everywhere anymore. So it goes for the active dropdown
menu items. Unfortunately by a mix of the default color values and the
custom ones, this leads to a white font on white background. This has
been fixed by extending the custom coloscheme for this specific
use-case.

Related to: #20