Replace default root README with custom version
This is a fork with adjustments for the Trustlines Foundation and their
blockchains. It is not a fork to work on features or bugs and open PRs
back to the origin. Also do links and shields (like CI builds) not refer
to the correct sources. All of the official documentation and more can
be found at their project.This should be reflected by the README
displayed to the user visiting this project.

This should not bother future upstream merges. The whole file can remain
as it is and must not care about conflicts or changes.

Related to: trustlines-network/project#798