Fix CSP issues for the Matomo statistics script
Since the Matomo script is fetched from our self hosted instance that
runs under a different domain, this was blocked on client side because
of the content security policy. The Blockscout does already have
a plugin for adding CSP headers. So this got extended to also work for
other domains like our stats server.

Related to: #32